weather/apps/users/serializers.py
from django.urls import path, include
from apps.users.models import User, City
from rest_framework import serializers


class CitySerializer(serializers.ModelSerializer):
    class Meta:
        model = City
        fields = ['city_id', 'city_name']

