weather/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.3
django-filter==21.1
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.11.2
jwcrypto==1.0
Markdown==3.3.6
mysqlclient==2.1.0
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
